refactor: 💡 blog and mypage
