+++ src/features/blog/components/blog-add-drawer.tsx
@@ -7,16 +7,8 @@ import {
 	DrawerHeader,
 	DrawerTitle,
 } from "@/components/ui/drawer";
-import { BlogAddForm } from "@/features/blog/components/blog-add-form";
-import type { Category } from "@prisma/client";
-import type { Dispatch, SetStateAction } from "react";
 
-type Props = {
-	categories: Omit<Category, "createdAt" | "updatedAt">[];
-	setOpen: Dispatch<SetStateAction<boolean>>;
-};
-
-export function BlogAddDrawer({ categories, setOpen }: Props) {
+export function BlogAddDrawer() {
 	return (
 		<>
 			<DrawerHeader>
@@ -25,16 +17,15 @@ export function BlogAddDrawer({ categories, setOpen }: Props) {
 					ブログに登録するデータを入力してください。
 				</DrawerDescription>
 			</DrawerHeader>
-			<BlogAddForm categories={categories} setDialogOpen={setOpen}>
-				<DrawerFooter>
-					<div className="grid grid-cols-2 gap-4">
-						<DrawerClose asChild>
-							<Button variant="outline">キャンセル</Button>
-						</DrawerClose>
-						<Button type="submit">保存</Button>
-					</div>
-				</DrawerFooter>
-			</BlogAddForm>
+			<div>XXXXXXXXXXXXXXXXX</div>
+			<DrawerFooter>
+				<div className="grid grid-cols-2 gap-4">
+					<DrawerClose asChild>
+						<Button variant="outline">キャンセル</Button>
+					</DrawerClose>
+					<Button type="submit">保存</Button>
+				</div>
+			</DrawerFooter>
 		</>
 	);
 }
