+++ src/features/blog/components/blog-add-form.tsx
@@ -14,19 +14,20 @@ import { ERROR_MESSAGES } from "@/constants";
 import { useToast } from "@/hooks/use-toast";
 import type { Category } from "@prisma/client";
 import { ClipboardPasteIcon } from "lucide-react";
-import type { Dispatch, ReactNode, SetStateAction } from "react";
 import { useEffect, useRef, useState } from "react";
 import { useSetRecoilState } from "recoil";
 import { addBlog } from "../actions/add-blog";
 import { blogContext } from "../stores/blog-context";
 
 type Props = {
-	children: ReactNode;
 	categories: Omit<Category, "createdAt" | "updatedAt">[];
-	setDialogOpen: Dispatch<SetStateAction<boolean>>;
 };
 
-export function BlogAddForm({ children, categories, setDialogOpen }: Props) {
+export function BlogAddForm({ categories }: Props) {
+	const titleInputRef = useRef<HTMLInputElement>(null);
+	const quoteInputRef = useRef<HTMLTextAreaElement>(null);
+	const urlInputRef = useRef<HTMLInputElement>(null);
+
 	const { toast } = useToast();
 
 	const [formData, setFormData] = useState<FormData>();
@@ -34,11 +35,6 @@ export function BlogAddForm({ children, categories, setDialogOpen }: Props) {
 
 	const NEW_CATEGORY_VALUE = "new"; // 新規の場合のcategoryのvalue
 	const [newCategoryInputOpen, setNewCategoryInputOpen] = useState(false);
-
-	const action = (_formData: FormData) => {
-		setDialogOpen(false);
-		setFormData(_formData);
-	};
 
 	useEffect(() => {
 		try {
@@ -61,6 +57,9 @@ export function BlogAddForm({ children, categories, setDialogOpen }: Props) {
 						description: state.message,
 					});
 					setFormData(undefined);
+					if (titleInputRef.current) titleInputRef.current.value = "";
+					if (quoteInputRef.current) quoteInputRef.current.value = "";
+					if (urlInputRef.current) urlInputRef.current.value = "";
 				};
 				submit();
 			}
@@ -77,15 +76,13 @@ export function BlogAddForm({ children, categories, setDialogOpen }: Props) {
 		setNewCategoryInputOpen(value === NEW_CATEGORY_VALUE);
 	};
 
-	const urlInputRef = useRef<HTMLInputElement>(null);
-
 	const handlePasteClick = async () => {
 		const clipboardText = await navigator.clipboard.readText();
 		if (urlInputRef.current !== null) urlInputRef.current.value = clipboardText;
 	};
 
 	return (
-		<form action={action} className="space-y-4 p-4">
+		<form action={setFormData} className="space-y-4 p-4">
 			<div className="space-y-1">
 				<Label htmlFor="category">カテゴリー</Label>
 				<Select
@@ -115,11 +112,11 @@ export function BlogAddForm({ children, categories, setDialogOpen }: Props) {
 			)}
 			<div className="space-y-1">
 				<Label htmlFor="title">タイトル</Label>
-				<Input id="title" name="title" required />
+				<Input id="title" name="title" ref={titleInputRef} required />
 			</div>
 			<div className="space-y-1">
 				<Label htmlFor="quote">ひとこと</Label>
-				<Textarea id="quote" name="quote" />
+				<Textarea id="quote" name="quote" ref={quoteInputRef} />
 			</div>
 			<div className="space-y-1">
 				<Label htmlFor="url">URL</Label>
@@ -130,7 +127,9 @@ export function BlogAddForm({ children, categories, setDialogOpen }: Props) {
 					</Button>
 				</div>
 			</div>
-			{children}
+			<Button type="submit" className="w-full">
+				保存
+			</Button>
 		</form>
 	);
 }
